+++ changelog.md
@@ -6,7 +6,6 @@ This changelog is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.
 
 ## Changed
 
-* Added node command `smart-contracts reserve init`
 * `setup-main-chain-state` command now uses native Rust to upsert the D-Parameter and upsert permissioned candidates
 
 ## Removed
@@ -20,6 +19,9 @@ This changelog is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.
 * Renamed of argument 'ogmios-host' to 'ogmios-url' in smart-contracts subcommands.
 
 ## Added
+
+* Command `smart-contracts reserve init`
+* Command `smart-contracts reserve create`
 
 # v1.4.0